refactored the js
